+++ oap-server/server-storage-plugin/storage-jdbc-hikaricp-plugin/src/main/java/org/apache/skywalking/oap/server/storage/plugin/jdbc/h2/dao/H2LogQueryDAO.java
@@ -162,13 +162,6 @@ public class H2LogQueryDAO implements ILogQueryDAO {
         Logs logs = new Logs();
         try (Connection connection = h2Client.getConnection()) {
 
-            try (ResultSet resultSet = h2Client.executeQuery(connection, buildCountStatement(sql.toString()), parameters
-                .toArray(new Object[0]))) {
-                while (resultSet.next()) {
-                    logs.setTotal(resultSet.getInt("total"));
-                }
-            }
-
             buildLimit(sql, from, limit);
 
             try (ResultSet resultSet = h2Client.executeQuery(
@@ -197,10 +190,6 @@ public class H2LogQueryDAO implements ILogQueryDAO {
         }
 
         return logs;
-    }
-
-    protected String buildCountStatement(String sql) {
-        return "select count(1) total from (select 1 " + sql + " )";
     }
 
     protected void buildLimit(StringBuilder sql, int from, int limit) {