+++ oap-server/server-storage-plugin/storage-jdbc-hikaricp-plugin/src/main/java/org/apache/skywalking/oap/server/storage/plugin/jdbc/postgresql/dao/PostgreSQLLogQueryDAO.java
@@ -29,9 +29,4 @@ public class PostgreSQLLogQueryDAO extends H2LogQueryDAO {
         super(h2Client, manager, maxSizeOfArrayColumn, numOfSearchValuesPerTag);
     }
     
-    @Override
-    protected String buildCountStatement(String sql) {
-        return "select count(*) total from (select 1 " + sql + " ) temp_table ";
-    }
-    
 }